// src/_proto/Protos/notifications.ts
export interface SetNotificationStatusRequest {
    /**
     * @generated from protobuf oneof: id
     */
    id: {
        oneofKind: "objectId";
        /**
         * @generated from protobuf field: string objectId = 1;
         */
        objectId: string;
    } | {
        oneofKind: "entityId";
        /**
         * @generated from protobuf field: int32 entityId = 3;
         */
        entityId: number;
    } | {
        oneofKind: undefined;
    };
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.NotificationItem.Status status = 2;
     */
    status: NotificationItem_Status;
}
// ...
export interface NotificationItem {
    /**
     * @generated from protobuf field: string id = 1;
     */
    id: string;
    /**
     * @generated from protobuf field: string objectId = 2;
     */
    objectId: string;
    /**
     * @generated from protobuf field: google.protobuf.Int32Value entityId = 3;
     */
    entityId?: Int32Value;
    /**
     * @generated from protobuf field: string type = 4;
     */
    type: string;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.NotificationItem.Status status = 5;
     */
    status: NotificationItem_Status;
    /**
     * @generated from protobuf field: google.protobuf.StringValue actor = 6;
     */
    actor?: StringValue;
    /**
     * @generated from protobuf field: google.protobuf.StringValue slug = 7;
     */
    slug?: StringValue;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp createdOn = 8;
     */
    createdOn?: Timestamp;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp timeRead = 9;
     */
    timeRead?: Timestamp;
    /**
     * @generated from protobuf field: map<string, string> data = 10;
     */
    data: {
        [key: string]: string;
    };
    /**
     * @generated from protobuf field: map<string, string> extraContent = 11;
     */
    extraContent: {
        [key: string]: string;
    };
}
/**
 * @generated from protobuf enum wuxiaworld.api.v2.NotificationItem.Status
 */
export enum NotificationItem_Status {
    /**
     * @generated from protobuf enum value: Unread = 0;
     */
    Unread = 0,
    /**
     * @generated from protobuf enum value: Read = 1;
     */
    Read = 1,
    /**
     * @generated from protobuf enum value: All = 2;
     */
    All = 2
}
// ...
class SetNotificationStatusRequest$Type extends MessageType<SetNotificationStatusRequest> {
    constructor() {
        super("wuxiaworld.api.v2.SetNotificationStatusRequest", [
            { no: 1, name: "objectId", kind: "scalar", oneof: "id", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "entityId", kind: "scalar", oneof: "id", T: 5 /*ScalarType.INT32*/ },
            { no: 2, name: "status", kind: "enum", T: () => ["wuxiaworld.api.v2.NotificationItem.Status", NotificationItem_Status] }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.SetNotificationStatusRequest
 */
export const SetNotificationStatusRequest = new SetNotificationStatusRequest$Type();
// ...
class NotificationItem$Type extends MessageType<NotificationItem> {
    constructor() {
        super("wuxiaworld.api.v2.NotificationItem", [
            { no: 1, name: "id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "objectId", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "entityId", kind: "message", T: () => Int32Value },
            { no: 4, name: "type", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 5, name: "status", kind: "enum", T: () => ["wuxiaworld.api.v2.NotificationItem.Status", NotificationItem_Status] },
            { no: 6, name: "actor", kind: "message", T: () => StringValue },
            { no: 7, name: "slug", kind: "message", T: () => StringValue },
            { no: 8, name: "createdOn", kind: "message", T: () => Timestamp },
            { no: 9, name: "timeRead", kind: "message", T: () => Timestamp },
            { no: 10, name: "data", kind: "map", K: 9 /*ScalarType.STRING*/, V: { kind: "scalar", T: 9 /*ScalarType.STRING*/ } },
            { no: 11, name: "extraContent", kind: "map", K: 9 /*ScalarType.STRING*/, V: { kind: "scalar", T: 9 /*ScalarType.STRING*/ } }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.NotificationItem
 */
export const NotificationItem = new NotificationItem$Type();